Refactor MTS API to decouple charities from base
