+++ src/lib/api/charity.ts
@@ -1,3 +1,5 @@
+import { API_BASE } from './mts';
+
 export type TCharityProgram = {
 	serviceName: string;
 	serviceCode: string;
@@ -15,19 +17,7 @@ export type TCharity = {
 	programs: TCharityProgram[];
 };
 
-export const API_BASE = 'https://test.mospay.mos.ru/mospaynew';
 export const FUNDS_ASSETS_BASE = API_BASE + '/newportal/assets/images/charity/';
-export const getAccessTokenUrl = (redirectUrl: string) => {
-	const AUTH_URL_BASE = 'https://login.mos.ru/sps/oauth/ae';
-	const params = new URLSearchParams({
-		scope: 'openid profile',
-		response_type: 'code',
-		redirect_uri: 'https://test.mospay.mos.ru/sudir/auth',
-		client_id: 'test.mospay.mos.ru',
-		state: redirectUrl,
-	});
-	return AUTH_URL_BASE + '?' + params.toString();
-};
 
 export const sampleCharities: TCharity[] = [
 	{
